Added processed to storage event
This is to make sure that the cron service is able to not repeat
proccessing over storage events that have already been processed.
While this doesn't cause an issue with resolving the event, it will
contribute to incorrect rates of reports being saved to the user,
in the case that an event is process more than one time.